Code cleanse
