+++ post_office/validators.py
@@ -1,28 +1,31 @@
+import re
+
 from django.core.exceptions import ValidationError
-from django.core.validators import validate_email
+from django.core.validators import validate_email as validate_plain_email, EmailValidator
 from django.template import Template, TemplateSyntaxError
 from django.utils.encoding import force_text
 
 from .compat import text_type
 
-
-def validate_email_with_name(value):
-    """
-    Validate email address.
+class FullEmailValidator(EmailValidator):
+    """ Simple validator that passes for email addresses bearing a display name
+    i.e. John Smith <john.smith@acme.gov>
 
     Both "Recipient Name <email@example.com>" and "email@example.com" are valid.
+
     """
-    value = force_text(value)
+    def __call__(self, value):
+        try:
+            res = super(FullEmailValidator, self).__call__(value)
+        except ValidationError:
+            try:
+                split_address = re.match(r'(.+) \<(.+@.+)\>', value)
+                display_name, email = split_address.groups()
+                super(FullEmailValidator, self).__call__(email)
+            except AttributeError:
+                raise ValidationError(self.message, code=self.code)
 
-    if '<' and '>' in value:
-        start = value.find('<') + 1
-        end = value.find('>')
-        if start < end:
-            recipient = value[start:end]
-    else:
-        recipient = value
-
-    validate_email(recipient)
+validate_email_with_name = FullEmailValidator(**dict(validate_plain_email.__dict__))
 
 
 def validate_comma_separated_emails(value):
